Fixed the stupid uninstall error
Currently, the driver's functionality doesn't really make sense (nor
does it really make sense). Maybe add an IOCTL to protect a process
after it is launched. (IOCTL takes in a pid as an input and add its to
the pid watch list). Currently the pid watch list doesn't really make
sense...lsass, csrss need to acquire handles to the process so it
effectively rejects a process from being created, rather than protect
it.